Split files over 1000 lines into modules under 700 lines
- Extract GLSL shaders from useWebGLOverlay.ts into hooks/webGLShaders.ts
  (buildVertexShader / buildFragmentShader functions)
- Extract app constants from App.tsx into appConfig.ts
  (shader groups, theme options, computeModuleHash, IS_PUBLIC_MODE)
- Extract 3D render block from App.tsx into components/App3DView.tsx
- Extract 2D render block from App.tsx into components/MainLayout.tsx
- Fix setPan prop: pass local state setter (not setLibPan) so pan syncs correctly
- Fix MainLayout shaderCatalog type to use ShaderMeta from storageApi
- Fix async handler types for handleLibrarySongLoad / onSyncLibrary
