+++ hooks/webGLShaders.ts
@@ -1,0 +1,544 @@
+// WebGL2 GLSL shader source builders for the WebGL overlay subsystem.
+// Parameterized by two runtime booleans that get baked into GLSL constants.
+
+export function buildVertexShader(useNoteSustainTailMode: boolean, isV021: boolean): string {
+  return `#version 300 es
+    precision highp float;
+    precision highp int;
+    precision highp usampler2D;
+
+    in vec2 a_pos;
+
+    // Flat varyings for fragment shader
+    flat out uvec2 v_cell;       // (packedA, packedB)
+    flat out vec4 v_chState0;    // (volume, pan, freq, trigger)
+    flat out vec4 v_chState1;    // (noteAge, activeEffect, effectValue, isMuted)
+    out vec2 v_uv;
+    flat out float v_active;     // playhead activation
+    flat out float v_hasNote;    // 1.0 if note data exists
+
+    uniform vec2 u_resolution;
+    uniform vec2 u_cellSize;
+    uniform vec2 u_offset;
+    uniform float u_cols;
+    uniform float u_rows;
+    uniform float u_playhead;
+    uniform int u_invertChannels;
+    uniform int u_layoutMode; // 1=Circ, 2=Horiz32, 3=Horiz64
+    uniform float u_innerRadius;
+    uniform float u_outerRadius;
+    uniform highp usampler2D u_cellData;    // RG32UI: (packedA, packedB)
+    uniform highp sampler2D u_channelState; // RGBA32F: row0=(vol,pan,freq,trig), row1=(age,eff,effVal,muted)
+
+    const float PI = 3.14159265359;
+    const float INNER_RADIUS = 0.3;
+    const float OUTER_RADIUS = 0.9;
+    const float CAP_SCALE_FACTOR = 0.88;
+    const uint NOTE_MIN = 1u;    // Minimum valid MIDI note
+    const uint NOTE_MAX = 96u;   // Maximum valid MIDI note
+    const uint NOTE_OFF = 97u;   // Note-off command
+    const uint NOTE_CUT = 98u;   // Note-cut command
+    const uint NOTE_FADE = 99u;  // Note-fade command
+    const bool USE_NOTE_SUSTAIN_TAIL_MODE = ${useNoteSustainTailMode ? 'true' : 'false'};
+    const bool IS_V021 = ${isV021 ? 'true' : 'false'};
+
+    void main() {
+        int id = gl_InstanceID;
+        int trackIndex = id % int(u_cols);
+        int stepIndex  = id / int(u_cols);
+
+        // Page-aware cell data fetch.
+        // Horizontal layouts page through the pattern (32 or 64 steps per page);
+        // circular layouts show all rows directly but must clamp to valid range.
+        int actualRow;
+        if (u_layoutMode == 2 || u_layoutMode == 3) {
+            float stepsPerPageH = (u_layoutMode == 3) ? 64.0 : 32.0;
+            int pageStart = int(floor(u_playhead / stepsPerPageH) * stepsPerPageH);
+            actualRow = clamp(pageStart + stepIndex, 0, int(u_rows) - 1);
+        } else {
+            actualRow = clamp(stepIndex, 0, int(u_rows) - 1);
+        }
+        uvec2 cellData = texelFetch(u_cellData, ivec2(trackIndex, actualRow), 0).rg;
+        v_cell = cellData;
+
+        // Fetch channel state (2 rows)
+        v_chState0 = texelFetch(u_channelState, ivec2(trackIndex, 0), 0);
+        v_chState1 = texelFetch(u_channelState, ivec2(trackIndex, 1), 0);
+
+        // Check for note data (quick presence detection — full unpacking in FS)
+        uint note = (cellData.r >> 24u) & 255u;
+        uint effCmd = (cellData.g >> 24u) & 255u;
+        uint volCmdFull = cellData.g & 255u; // Low byte of packedB = volCmdFull
+        // Duration data for sustain detection in VS
+        // packedA layout: [note:8][inst:8][duration:8][volPacked:8]
+        uint durationRaw = (cellData.r >> 8u) & 255u;
+        // packedB layout: [effCmd:8][effVal:8][durationFlags:7][reserved:1][volCmd:8]
+        // durationFlags: [rowOffset:6][isNoteOff:1]
+        uint durationFlags = (cellData.g >> 8u) & 127u;
+        uint rowOffset = (durationFlags >> 1u) & 63u;
+        bool isNoteOffFlag = (durationFlags & 1u) != 0u;
+        // A cell is visible if it has a note, expression data, OR is part of an active sustain
+        bool hasNoteOrExpr = (note > 0u) || (effCmd > 0u) || (volCmdFull > 0u);
+        bool isSustainCell = (note >= NOTE_MIN && note <= NOTE_MAX) && (durationRaw > 1u) && (rowOffset > 0u) && !isNoteOffFlag;
+        v_hasNote = (hasNoteOrExpr || isSustainCell) ? 1.0 : 0.0;
+
+        // Playhead Logic — Directional Sustain/Duration Tail
+        // Unpack duration from packedA: [note:8][inst:8][duration:8][volPacked:8]
+        float noteDuration = (durationRaw > 0u) ? float(durationRaw) :
+                             ((note >= NOTE_MIN && note <= NOTE_MAX) ? 1.5 : 0.0);
+        bool hasPitchNote = (note >= NOTE_MIN && note <= NOTE_MAX);
+        bool isNoteOffCmd = (note == NOTE_OFF || note == NOTE_CUT || note == NOTE_FADE);
+        // Keep this predicate in sync with FS + patternv0.45b.wgsl sustain logic.
+        bool isNoteOnCell = hasPitchNote && (rowOffset == 0u) && !isNoteOffFlag && !isNoteOffCmd;
+
+        float activation = 0.0;
+        if (IS_V021) {
+            // V0.21 Strict Delta: absolute row-based with pattern-loop wrap (64-row patterns)
+            float delta = u_playhead - float(actualRow);
+            if (delta < -32.0) delta += 64.0;
+            else if (delta > 32.0) delta -= 64.0;
+
+            // For v0.21, any cell with data gets at least a 1-row duration so expression-only steps light up
+            float cellDuration = (noteDuration > 0.0) ? noteDuration :
+                                 ((note > 0u || volCmdFull > 0u || effCmd > 0u) ? 1.0 : 0.0);
+
+            bool isActive = (delta >= 0.0 && delta < cellDuration);
+            // Note-off rows cut the light immediately
+            if (isNoteOffCmd || isNoteOffFlag) isActive = false;
+
+            if (isActive) {
+                if (note > 0u) {
+                    activation = 2.0; // Bright note (Blue/Cyan)
+                } else if (volCmdFull > 0u || effCmd > 0u) {
+                    activation = 1.0; // Dim expression (Amber/Orange)
+                }
+            }
+        } else {
+            float stepsPerPage = (u_layoutMode == 3) ? 64.0 : 32.0;
+            float relativePlayhead = mod(u_playhead, stepsPerPage);
+            float delta = relativePlayhead - float(stepIndex);
+            if (delta < -stepsPerPage / 2.0) delta += stepsPerPage;
+            if (delta > stepsPerPage / 2.0) delta -= stepsPerPage;
+
+            // v_active = 1.0 only when playhead is within the note's duration window
+            // Fast fade-out in the last 0.5 steps so it doesn't snap off harshly
+            bool sustainCellFilter = USE_NOTE_SUSTAIN_TAIL_MODE ? isNoteOnCell : hasPitchNote;
+            if (sustainCellFilter && noteDuration > 0.0 && delta >= 0.0 && delta <= noteDuration) {
+                activation = 1.0;
+                float fadeEnd = noteDuration - 0.5;
+                if (delta > fadeEnd && fadeEnd > 0.0) {
+                    activation = 1.0 - smoothstep(fadeEnd, noteDuration, delta);
+                }
+            }
+        }
+        v_active = activation;
+
+        // Positioning Logic
+        if (u_layoutMode == 2 || u_layoutMode == 3) {
+            // --- HORIZONTAL LAYOUT (32-step or 64-step) ---
+            float capScale = min(u_cellSize.x, u_cellSize.y) * CAP_SCALE_FACTOR;
+            if (v_hasNote < 0.5) capScale = 0.0;
+            capScale *= 1.0 + (0.2 * activation);
+
+            float cellX = u_offset.x + float(stepIndex)  * u_cellSize.x;
+            float cellY = u_offset.y + float(trackIndex) * u_cellSize.y;
+
+            vec2 centered = a_pos * capScale + vec2(cellX + u_cellSize.x * 0.5, cellY + u_cellSize.y * 0.5);
+            vec2 ndc = (centered / u_resolution) * 2.0 - 1.0;
+            ndc.y = -ndc.y;
+            gl_Position = vec4(ndc, 0.0, 1.0);
+
+        } else {
+            // --- CIRCULAR LAYOUT ---
+            float numTracks = u_cols;
+            float trackIndexF = float(trackIndex);
+            if (u_invertChannels == 0) { trackIndexF = numTracks - 1.0 - trackIndexF; }
+
+            // Match WGSL: no floor() — keeps sub-pixel parity with the GPU grid
+            float maxRadius = u_outerRadius;
+            float minRadius = u_innerRadius;
+            float ringDepth = (maxRadius - minRadius) / numTracks;
+
+            // Match WGSL: cell center is at ring-start, not ring-center
+            float pixelRadius = minRadius + trackIndexF * ringDepth;
+
+            float totalSteps = 64.0;
+            float anglePerStep = (2.0 * PI) / totalSteps;
+            float theta = -1.570796 + float(stepIndex) * anglePerStep;
+
+            float circumference = 2.0 * PI * pixelRadius;
+            float arcLength = circumference / totalSteps;
+            float btnW = arcLength * CAP_SCALE_FACTOR;
+            float btnH = ringDepth * 0.92;
+
+            float circPlayhead = mod(u_playhead, totalSteps);
+            float circDelta = circPlayhead - float(stepIndex);
+            if (circDelta < -totalSteps / 2.0) circDelta += totalSteps;
+
+            float circDuration = (durationRaw > 0u) ? float(durationRaw) :
+                                 ((note >= NOTE_MIN && note <= NOTE_MAX) ? 1.5 : 0.0);
+            float circActivation = 0.0;
+            if (circDuration > 0.0 && circDelta >= 0.0 && circDelta <= circDuration) {
+                circActivation = 1.0;
+                float circFadeEnd = circDuration - 0.5;
+                if (circDelta > circFadeEnd && circFadeEnd > 0.0) {
+                    circActivation = 1.0 - smoothstep(circFadeEnd, circDuration, circDelta);
+                }
+            }
+            float popScale = (v_hasNote > 0.5) ? (1.0 + 0.2 * circActivation) : 0.0;
+            btnW *= popScale;
+            btnH *= popScale;
+            v_active = circActivation;
+
+            vec2 localPos = a_pos * vec2(btnW, btnH);
+            float rotAng = theta + 1.570796;
+            float cA = cos(rotAng); float sA = sin(rotAng);
+            float rotX = localPos.x * cA - localPos.y * sA;
+            float rotY = localPos.x * sA + localPos.y * cA;
+
+            vec2 center = u_resolution * 0.5;
+            float worldX = center.x + cos(theta) * pixelRadius + rotX;
+            float worldY = center.y + sin(theta) * pixelRadius + rotY;
+
+            vec2 ndc = vec2(
+                (worldX / u_resolution.x) * 2.0 - 1.0,
+                1.0 - (worldY / u_resolution.y) * 2.0
+            );
+            gl_Position = vec4(ndc, 0.0, 1.0);
+        }
+
+        v_uv = a_pos + 0.5;
+    }
+    `;
+}
+
+export function buildFragmentShader(useNoteSustainTailMode: boolean, isV021: boolean): string {
+  return `#version 300 es
+    precision highp float;
+    precision highp int;
+
+    flat in uvec2 v_cell;
+    flat in vec4 v_chState0;
+    flat in vec4 v_chState1;
+    in vec2 v_uv;
+    flat in float v_active;
+    flat in float v_hasNote;
+
+    // Note range constants (must match VS)
+    const uint NOTE_MIN = 1u;
+    const uint NOTE_MAX = 96u;
+    const uint NOTE_OFF = 97u;
+    const uint NOTE_CUT = 98u;
+    const uint NOTE_FADE = 99u;
+    const bool USE_NOTE_SUSTAIN_TAIL_MODE = ${useNoteSustainTailMode ? 'true' : 'false'};
+    const bool IS_V021 = ${isV021 ? 'true' : 'false'};
+
+    uniform sampler2D u_capTexture;
+    uniform float u_bloomIntensity;
+    uniform float u_timeSec;
+
+    out vec4 fragColor;
+
+    // --- Pitch-to-color: cosine palette matching WGSL shaders (Bug 4 fix) ---
+    vec3 neonPalette(float t) {
+        vec3 a = vec3(0.5, 0.5, 0.5);
+        vec3 b = vec3(0.5, 0.5, 0.5);
+        vec3 c = vec3(1.0, 1.0, 1.0);
+        vec3 d = vec3(0.0, 0.33, 0.67);
+        return a + b * cos(6.28318 * (c * t + d));
+    }
+
+    float pitchClassFromIndex(uint note) {
+        if (note < 1u || note > 96u) return 0.0;
+        uint semi = (note - 1u) % 12u;
+        return float(semi) / 12.0;
+    }
+
+    // --- SDF: Rounded Box ---
+    float sdRoundedBox(vec2 p, vec2 b, float r) {
+        vec2 q = abs(p) - b + vec2(r);
+        return length(max(q, vec2(0.0))) + min(max(q.x, q.y), 0.0) - r;
+    }
+
+    // --- Single LED emitter diode ---
+    vec4 drawEmitterDiode(vec2 uv, float intensity, vec3 color, bool isOn) {
+        vec2 diodeSize = vec2(0.28, 0.14);
+        float dDiode = sdRoundedBox(uv, diodeSize * 0.5, 0.06);
+        vec2 dieSize = vec2(0.10, 0.05);
+        float dDie = sdRoundedBox(uv, dieSize * 0.5, 0.02);
+        float diodeMask = 1.0 - smoothstep(0.0, 0.015, dDiode);
+        float dieMask = 1.0 - smoothstep(0.0, 0.008, dDie);
+        vec3 diodeColor = vec3(0.06, 0.06, 0.08);
+        if (isOn) {
+            // Central hotspot: tight Gaussian emitting directly under the cap
+            // vec2(0.06, 0.03) = die half-size in UV space; 2.5 = falloff sharpness
+            float dist = length(uv / vec2(0.06, 0.03));
+            float hotspot = exp(-dist * 2.5) * intensity;
+
+            // Subsurface scatter: wider bleed through frosted plastic
+            // 0.4 = scatter radius multiplier; 0.3 = max scatter brightness
+            float scatter = smoothstep(1.0, 0.0, dist * 0.4) * 0.3 * intensity;
+
+            vec3 dieGlow = color * (1.0 + intensity * 4.0);
+            vec3 housingGlow = color * 0.12 * intensity;
+            diodeColor = mix(housingGlow, dieGlow, dieMask);
+            diodeColor += color * (hotspot * 0.6 + scatter * 0.4);
+        }
+        return vec4(diodeColor, diodeMask);
+    }
+
+    // --- Unified three-emitter lens cap ---
+    vec4 drawUnifiedLensCap(
+        vec2 uv, vec2 lensSize,
+        vec4 topEmitter, vec4 midEmitter, vec4 botEmitter,
+        float aa
+    ) {
+        vec2 p = uv;
+        float dBox = sdRoundedBox(p, lensSize * 0.5, 0.12);
+        if (dBox > 0.0) return vec4(0.0);
+
+        vec2 topPos = vec2(0.0, -0.28);
+        vec2 midPos = vec2(0.0, 0.0);
+        vec2 botPos = vec2(0.0, 0.28);
+
+        float radial = length(p / (lensSize * 0.5));
+        float edgeThickness = 0.18 + radial * 0.12;
+        float centerThickness = 0.06;
+
+        vec3 n = normalize(vec3(p.x * 2.5 / lensSize.x, p.y * 2.5 / lensSize.y, 0.35));
+        vec3 viewDir = vec3(0.0, 0.0, 1.0);
+        float fresnel = pow(1.0 - abs(dot(n, viewDir)), 2.5);
+
+        vec4 topDiode = drawEmitterDiode(uv - topPos, topEmitter.a, topEmitter.rgb, topEmitter.a > 0.05);
+        vec4 midDiode = drawEmitterDiode(uv - midPos, midEmitter.a, midEmitter.rgb, midEmitter.a > 0.05);
+        vec4 botDiode = drawEmitterDiode(uv - botPos, botEmitter.a, botEmitter.rgb, botEmitter.a > 0.05);
+
+        vec3 combinedDiode = vec3(0.06, 0.06, 0.08);
+        if (botDiode.a > 0.0) combinedDiode = mix(combinedDiode, botDiode.rgb, botDiode.a);
+        if (midDiode.a > 0.0) combinedDiode = mix(combinedDiode, midDiode.rgb, midDiode.a);
+        if (topDiode.a > 0.0) combinedDiode = mix(combinedDiode, topDiode.rgb, topDiode.a);
+        float diodeMask = max(max(topDiode.a, midDiode.a), botDiode.a);
+
+        float refractionStrength = (1.0 - radial * 0.6) * 0.04;
+        vec2 refractOffset = p * refractionStrength;
+
+        // Tightened subsurface scattering (matching v0.50 §1)
+        vec3 subsurfaceGlow = vec3(0.0);
+        float distTop = length(uv - topPos - refractOffset * 0.3);
+        float scatterTop = exp(-distTop * 9.0) * topEmitter.a;
+        subsurfaceGlow += topEmitter.rgb * scatterTop * 2.2;
+        float distMid = length(uv - midPos - refractOffset * 0.5);
+        float scatterMid = exp(-distMid * 7.5) * midEmitter.a;
+        subsurfaceGlow += midEmitter.rgb * scatterMid * 3.0;
+        float distBot = length(uv - botPos - refractOffset * 0.3);
+        float scatterBot = exp(-distBot * 9.0) * botEmitter.a;
+        subsurfaceGlow += botEmitter.rgb * scatterBot * 2.2;
+        // Per-emitter fringe (replaces shared diffusion)
+        subsurfaceGlow += topEmitter.rgb * exp(-distTop * 6.0) * topEmitter.a * 0.15;
+        subsurfaceGlow += midEmitter.rgb * exp(-distMid * 6.0) * midEmitter.a * 0.15;
+        subsurfaceGlow += botEmitter.rgb * exp(-distBot * 6.0) * botEmitter.a * 0.15;
+
+        // Glass base color
+        vec3 bgColor = vec3(0.04, 0.04, 0.05);
+        vec3 activeColor = midEmitter.rgb * midEmitter.a;
+        activeColor = mix(activeColor, topEmitter.rgb, topEmitter.a * 0.5);
+        activeColor = mix(activeColor, botEmitter.rgb, botEmitter.a * 0.5);
+        float totalGlow = topEmitter.a + midEmitter.a + botEmitter.a;
+        vec3 litTint = mix(vec3(0.92, 0.93, 0.98), activeColor, min(totalGlow * 0.4, 0.4));
+        vec3 glassBaseColor = mix(bgColor * 0.12, litTint, 0.88);
+
+        float edgeAlpha = smoothstep(0.0, aa * 2.0, -dBox);
+        float diodeVisibility = diodeMask * 0.55;
+        float baseAlpha = 0.72 + 0.28 * fresnel;
+        float emitterLift = clamp(topEmitter.a * 0.4 + botEmitter.a * 0.4, 0.0, 0.7);
+        float alpha = mix(baseAlpha, 0.32, min(diodeVisibility + emitterLift, 0.9)) * edgeAlpha;
+
+        vec3 lightDir = vec3(0.4, -0.7, 0.6);
+        float diff = max(0.0, dot(n, normalize(lightDir)));
+        float spec = pow(max(0.0, dot(reflect(-normalize(lightDir), n), viewDir)), 40.0);
+        vec3 litGlassColor = glassBaseColor * (0.45 + 0.55 * diff) + vec3(spec * 0.25);
+
+        vec3 finalColor = bgColor;
+        float diodeBlend = diodeMask * (1.0 - alpha * 0.65);
+        finalColor = mix(finalColor, combinedDiode, diodeBlend);
+        finalColor = mix(finalColor, litGlassColor, alpha);
+        finalColor += subsurfaceGlow * 1.8;
+
+        // Concentrated glow (tightened radii matching v0.50 §1)
+        if (midEmitter.a > 0.05) {
+            float midGlowDist = length(uv - midPos - refractOffset * 0.5);
+            float midGlow = (1.0 - smoothstep(0.0, 0.18, midGlowDist)) * midEmitter.a * 0.5;
+            finalColor += midEmitter.rgb * midGlow;
+        }
+        if (topEmitter.a > 0.05) {
+            float topGlowDist = length(uv - topPos - refractOffset * 0.3);
+            float topGlow = (1.0 - smoothstep(0.0, 0.14, topGlowDist)) * topEmitter.a * 0.3;
+            finalColor += topEmitter.rgb * topGlow;
+        }
+        if (botEmitter.a > 0.05) {
+            float botGlowDist = length(uv - botPos - refractOffset * 0.3);
+            float botGlow = (1.0 - smoothstep(0.0, 0.14, botGlowDist)) * botEmitter.a * 0.3;
+            finalColor += botEmitter.rgb * botGlow;
+        }
+
+        finalColor += fresnel * vec3(0.9, 0.95, 1.0) * 0.18 * (1.0 + radial * 0.5);
+
+        // Horizontal separator shadows
+        float sepShadowTop = (1.0 - smoothstep(0.0, 0.015, abs(p.y - (-0.14)))) * 0.35;
+        float sepShadowBot = (1.0 - smoothstep(0.0, 0.015, abs(p.y - 0.14))) * 0.35;
+        finalColor -= finalColor * (sepShadowTop + sepShadowBot);
+
+        float vignette = 1.0 - radial * radial * 0.25;
+        finalColor *= vignette;
+
+        return vec4(finalColor, edgeAlpha);
+    }
+
+    void main() {
+        // Unpack cell data
+        uint packedA = v_cell.x;
+        uint packedB = v_cell.y;
+
+        uint note = (packedA >> 24u) & 255u;
+        uint instRaw = (packedA >> 16u) & 255u;
+        uint durationRaw = (packedA >> 8u) & 255u;
+        uint volPacked = packedA & 255u;
+        uint effCmd = (packedB >> 24u) & 255u;
+        uint effVal = (packedB >> 16u) & 255u;
+        uint durationFlags = (packedB >> 8u) & 127u;
+        uint volCmdFull = packedB & 255u;
+
+        bool isExpressionOnly = (instRaw & 128u) != 0u;
+        uint volCmd = (volPacked >> 4u) << 4u;
+
+        // Note state detection
+        bool hasNote = (note >= NOTE_MIN && note <= NOTE_MAX);
+        bool isNoteOff = (note == NOTE_OFF || note == NOTE_CUT || note == NOTE_FADE);
+        bool hasExpression = (volCmd > 0u) || (effCmd > 0u) || (volCmdFull > 0u);
+
+        // Duration / sustain state (DURA data packed by calculateNoteDurations)
+        // durationFlags layout: [rowOffset:6][isNoteOff:1]
+        uint rowOffset = (durationFlags >> 1u) & 63u;
+        bool isNoteOffFlag = (durationFlags & 1u) != 0u;
+        bool isSustaining = hasNote && (durationRaw > 1u) && (rowOffset > 0u) && !isNoteOffFlag;
+        bool isNoteOnRow = hasNote && (rowOffset == 0u);
+        // 0.0 at note start → 1.0 at note end; used to fade trail glow
+        float sustainProgress = (durationRaw > 1u) ? float(rowOffset) / float(durationRaw) : 0.0;
+
+        // Channel state
+        float chVolume = v_chState0.x;
+        float chTrigger = v_chState0.w;
+        float chNoteAge = v_chState1.x;
+        float chIsMuted = v_chState1.w;
+        bool isMuted = chIsMuted > 0.5;
+
+        float bloom = u_bloomIntensity;
+        float aa = fwidth(v_uv.y) * 0.33;
+
+        // Map UV to lens-cap space (-0.5..0.5)
+        vec2 lensUV = v_uv - vec2(0.5);
+
+        // --- THREE EMITTER SETUP ---
+
+        vec3 topColor = vec3(0.15, 0.5, 1.0) * (1.5 + bloom * 2.0);
+        vec3 midColor = vec3(0.15);
+        vec3 botColor = vec3(1.0, 0.55, 0.1) * (1.5 + bloom * 2.0);
+        float topIntensity = 0.0;
+        float midIntensity = 0.12;
+        float botIntensity = 0.0;
+
+        if (IS_V021) {
+            // V0.21 Strict Two-State LED Logic
+            // v_active: 2.0 = note active, 1.0 = expression active, 0.0 = off
+            if (!isMuted) {
+                if (v_active > 1.5) {
+                    // Bright Blue/Cyan glow for active note
+                    topIntensity = 3.0;
+                    float pitchHue = pitchClassFromIndex(note);
+                    midColor = neonPalette(pitchHue);
+                    midIntensity = 0.6 + bloom * 2.0;
+                    botIntensity = 0.6;
+                } else if (v_active > 0.5) {
+                    // Dim Amber/Orange glow for active expression (no note)
+                    botIntensity = 1.2;
+                }
+            }
+        } else {
+            // EMITTER 1 (TOP): Blue/Cyan — Note trigger + sustain trail
+            vec3 blueColor = vec3(0.15, 0.5, 1.0);
+            topIntensity = 0.0;
+            if (!isMuted) {
+                if (chTrigger > 0.5 && isNoteOnRow) {
+                    // Bright flash on note-on trigger row
+                    topIntensity = 3.0;
+                } else if (isNoteOnRow) {
+                    // Note-on row near playhead
+                    topIntensity = v_active * 0.8;
+                } else if (isSustaining && !USE_NOTE_SUSTAIN_TAIL_MODE) {
+                    // Sustain trail: fades from 0.4 (note start) to 0.2 (note end)
+                    // Clamped to 0.15 minimum to keep a dim glow visible
+                    float sustainFade = 0.4 - sustainProgress * 0.2;
+                    topIntensity = max(sustainFade, 0.15);
+                } else if (USE_NOTE_SUSTAIN_TAIL_MODE && isNoteOnRow && v_active > 0.0) {
+                    topIntensity = max(topIntensity, v_active * 0.8);
+                } else if (isNoteOff || isNoteOffFlag) {
+                    // Note-off: completely dark (channel was cut)
+                    topIntensity = 0.0;
+                }
+            }
+            topColor = blueColor * (1.5 + bloom * 2.0);
+
+            // EMITTER 2 (MIDDLE): Pitch-colored note indicator
+            vec3 noteColor = vec3(0.15);
+            midIntensity = 0.12;
+            if (hasNote && !isExpressionOnly) {
+                float pitchHue = pitchClassFromIndex(note);
+                noteColor = neonPalette(pitchHue);
+                if (isNoteOnRow) {
+                    // Strong glow for note-on
+                    midIntensity = 0.6 + bloom * 2.0;
+                    if (USE_NOTE_SUSTAIN_TAIL_MODE && v_active > 0.0) {
+                        midIntensity = max(midIntensity, 0.6 + v_active * (0.8 + bloom));
+                    }
+                } else if (isSustaining && !USE_NOTE_SUSTAIN_TAIL_MODE) {
+                    // Sustain: fades from 0.5 (note start) to 0.25 (note end)
+                    // Clamped to 0.15 minimum for a dim ringing glow
+                    float sustainBright = 0.5 - sustainProgress * 0.25;
+                    midIntensity = max(sustainBright, 0.15);
+                }
+                if (isMuted) midIntensity *= 0.3;
+            }
+            midColor = noteColor;
+
+            // EMITTER 3 (BOTTOM): Amber/Orange — Expression indicator
+            vec3 amberColor = vec3(1.0, 0.55, 0.1);
+            botIntensity = 0.0;
+            if (!isMuted) {
+                if (isExpressionOnly && hasExpression) {
+                    // Expression-only step (no note trigger): distinct amber glow
+                    botIntensity = 2.0;
+                } else if (hasExpression && hasNote) {
+                    // Note + expression: moderate amber
+                    botIntensity = 1.2;
+                } else if (hasNote && v_active > 0.5) {
+                    botIntensity = 0.6;
+                }
+            }
+            botColor = amberColor * (1.5 + bloom * 2.0);
+        }
+
+        // Draw unified lens cap
+        vec2 lensSize = vec2(0.6, 0.82);
+        vec4 lens = drawUnifiedLensCap(
+            lensUV, lensSize,
+            vec4(topColor, topIntensity),
+            vec4(midColor, midIntensity),
+            vec4(botColor, botIntensity),
+            aa
+        );
+
+        fragColor = vec4(lens.rgb, lens.a * 0.95);
+        if (fragColor.a < 0.01) discard;
+    }
+    `;
+}
